Splits manuallyBuiltDeps.js in modules for better traceability and concat them in the proper order
